instances-aware Cassandra restart
Adds support of rolling Cassandra restarts on hosts that have multiple
Cassandra instances running.  Each such instance is specified using the
instance-specific hostname, and had an `instance` host variable indicating
the instance id.

This is a little janky.  The way it is implemented, you will see output
for two different forms of restart, one for the legacy method, and another
that is instance-based.  Depending on the node type, one will be skipped
and the other ran.

Bug: https://phabricator.wikimedia.org/T117114

not pretty, but functionally working